Added a starter /event/* page
I have a feeling that Caleb is not going to like my changes to so many files...
I apologize in advance.

Anyway, there are two matches artificially added for debugging/viewing purposes. Feel free to remove them when the backend has correctly implemented the endpoint.

Also I styled "<a>" so that it did not look so awful. Hopefully that was alright.