Unstyled Match Page
Until the backend is update to meet the TODOs listed in src/routes/match/index.js
there is no more to be added (as far as I can tell).

I have very basic styling in order to verify that my output data was working.

Also the images are not loading and I can not figure out why, the webpack config file
is very complicated.